refactor: structure fsd and move routes
