+++ src/app/providers/router/index.ts
@@ -1,0 +1,3 @@
+import AppRouter from "./ui/AppRouter";
+
+export { AppRouter };
