+++ src/index.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
-import App from "./App";
+import App from "./app/App";
 import { BrowserRouter } from "react-router-dom";
 
 const root = createRoot(document.getElementById("root"));
