+++ src/pages/About/index.ts
@@ -1,0 +1,3 @@
+import { AboutLazy } from "./ui/AboutLazy";
+
+export { AboutLazy as About };
