+++ src/pages/Main/index.ts
@@ -1,0 +1,3 @@
+import { MainLazy } from "./ui/MainLazy";
+
+export { MainLazy as Main };
