+++ src/pages/Main/ui/MainLazy.tsx
